Bunch of changes
Added no_optimize_begin, no_optimize_end
Fixed issues in process.h
Moved string_from into migi_core
Added memory_poison functions from ASan
Removed arena_push_nonzero
arena_push now takes the optional zeroed parameter (defaults to true)
renamed the old arena_push_bytes to arena_push_bytes_opt
arena_push_bytes now takes the same ArenaOpt structure as an optional parameter
Minor fixes

+++ src/cli_parse.h
@@ -523,7 +523,7 @@ static bool cli__parse_value(Arena *arena, CliArg *cli_arg, Str key, Str value, 
                 strlist_extend(&cli_arg->as_list, &items);
                 cli_arg->was_set = true;
             }
-            cli_arg->found_args += items.length;
+            cli_arg->found_args += (int32_t)items.length;
         } break;
         default:
             migi_unreachable();
